first working driver for bcm2835 & bcm2711 controllers
The main function probes the hardware with for each implemented
controller. If found, all vbus device resources are acquired and the
controller is enabled.

With the controller info, main starts the server frontend. Via the
factory interface new client interfaces are created. The I2C device
address must be passed to the factory.
Errors during the create call are either that the device does not exists
or it is already used by another client.
Multiplexing an I2C device amond different clients is not supported.

This also extends the frontend library where needed.

Change-Id: I8f036eabe2e30e636d8408623036923b2e159a88