introduce virtio-i2c-device as second createable device interface
The virtio-i2c-device template needs a request handler to implement
the required read/write operations. The new request handler translates
the read/write requests to the controller interface and sends them
along for the controller to send on the I2C bus.

This extends the valid parameter range for the type parameter of the
factory create interface to [0,1].
Type 0 creates a virtio-i2c-device and type 1 a plain i2c-device
implementing the RPC interface. Both need the hardware device address
as second parameter.

Change-Id: Ia0cbac095e6ad010f156d832764d0ba42ca4da72