Wait for controller to flag start of the transaction
The controller is slower to initialize a new transaction than the driver
is with executing the code. The driver needs to wait for the controller
to flag the start of the transaction before proceeding with read or
write.

We now actively poll the Status.TA flag and if necessary wait for even
more time until we abort and try to bring the controller into a usable
state again.

Change-Id: Icc7eda49a7638eb506a1ed3ec7a8ac0b869ee621